+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.7.14] - Unreleased
 ### Added
-- YAML tags and namespace to support DIFI 1.2 deviations to reference level and buffer size context fields
+  - YAML tags and namespace to support DIFI 1.2 deviations to reference level and buffer size context fields
+  - Support for packet types in multiple information classes
+  - 'any' type for tsm fields
+    - Adds additional getter/setter functions for tsm
+    - Removes explicit matching on tsm
 ### Changed
 - Make Class ID field `pad_bits` runtime dynamic
   - Do not check value in match function