+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.14] - Unreleased
+## [0.7.14] - 2024-11-06
 ### Added
   - YAML tags and namespace to support DIFI 1.2 deviations to reference level and buffer size context fields
   - Support for packet types in multiple information classes